+++ 조이-이지은/react/react-kurly/src/App.jsx
@@ -1,10 +1,19 @@
 import React from "react";
-import Home from "./components/r-home";
+import Home from "./components/pages/Home";
+import Login from "./components/pages/Login";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+import Header from "./components/layout/Header";
 function App() {
   return (
-    <div>
-      <Home />
-    </div>
+    <BrowserRouter>
+      <div>
+        <Header />
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/login" element={<Login />} />
+        </Routes>
+      </div>
+    </BrowserRouter>
   );
 }
