+++ 조이-이지은/react/react-kurly/src/components/styles/styles.css
@@ -318,3 +318,7 @@ body {
   align-items: center;
   margin-bottom: 30rem;
 }
+
+img {
+  cursor: pointer;
+}
